Updating molecule execution for more scenarios
