--- roles/splunk_heavy_forwarder/molecule/manual-forwards/INSTALL.rst
@@ -1,16 +1,0 @@
-*******
-Docker driver installation guide
-*******
-
-Requirements
-============
-
-* General molecule dependencies (see https://molecule.readthedocs.io/en/latest/installation.html)
-* Docker Engine
-* docker-py
-* docker
-
-Install
-=======
-
-    $ sudo pip install docker-py
