+++ roles/splunk_heavy_forwarder/molecule/manual-forwards/group_vars/all.yml
@@ -1,0 +1,10 @@
+---
+splunk:
+  role: splunk_heavy_forwarder
+  build_location: https://download.splunk.com/products/splunk/releases/7.3.1.1/linux/splunk-7.3.1.1-7651b7244cf2-Linux-x86_64.tgz
+  forward_servers:
+    - forwarder1.test.com
+    - forwarder2.test.com
+  hec:
+    token: abcd1234
+  password: helloworld
