+++ roles/splunk_heavy_forwarder/molecule/manual-forwards/molecule.yml
@@ -4,17 +4,45 @@ dependency:
 driver:
   name: docker
 platforms:
-  - name: instance
-    image: splunk/splunk
-    command: no-provision
+  - name: splunk-hf-centos8
+    image: geerlingguy/docker-centos8-ansible:latest
+    pre_build_image: true
+    command: "/lib/systemd/systemd"
+    volumes:
+      - "/opt/splunk/var"
+    env:
+      http_proxy: "${http_proxy}"
+      https_proxy: "${https_proxy}"
+      no_proxy: "${no_proxy}"
+  - name: splunk-hf-debian10
+    image: geerlingguy/docker-debian10-ansible:latest
+    pre_build_image: true
+    command: "/lib/systemd/systemd"
+    volumes:
+      - "/opt/splunk/var"
+    env:
+      http_proxy: "${http_proxy}"
+      https_proxy: "${https_proxy}"
+      no_proxy: "${no_proxy}"
+  - name: splunk-hf-redhat8
+    image: geerlingguy/docker-ubi8-ansible:latest
+    pre_build_image: true
+    command: "/lib/systemd/systemd"
+    volumes:
+      - "/opt/splunk/var"
     env:
       http_proxy: "${http_proxy}"
       https_proxy: "${https_proxy}"
       no_proxy: "${no_proxy}"
 provisioner:
   name: ansible
-  log: true
+  env:
+    ANSIBLE_HASH_BEHAVIOUR: merge
+  inventory:
+    links:
+      hosts: ../hosts.yml
+      group_vars: ./group_vars/
 verifier:
   name: testinfra
   options:
-    junit-xml: ../../../../tests/results/heavy-manual-forwards-results.xml
+    junit-xml: ../../../../tests/results/hf-manual-forwards-results.xml
