+++ roles/splunk_heavy_forwarder/molecule/manual-forwards/prepare.yml
@@ -16,12 +16,3 @@
           - curl
           - net-tools
         state: latest
-
-    - name: Setup yum proxy
-      lineinfile:
-        path: /etc/yum.conf
-        regexp: '^proxy='
-        line: "proxy={{ lookup('env','http_proxy') }}"
-      when:
-        - "lookup('env','http_proxy') is defined"
-        - "lookup('env','http_proxy')  | length > 0"
