+++ roles/splunk_standalone/molecule/systemd/molecule.yml
@@ -13,6 +13,7 @@ platforms:
       - /sys/fs/cgroup/memory/system.slice/Splunkd.service
       - /sys/fs/cgroup/cpu/system.slice/Splunkd.service
     volumes:
+      - "/opt/splunk/var"
       - "/sys/fs/cgroup:/sys/fs/cgroup:rw"
     command: "/lib/systemd/systemd"
     env:
@@ -21,8 +22,13 @@ platforms:
       no_proxy: "${no_proxy}"
 provisioner:
   name: ansible
-  log: true
+  env:
+    ANSIBLE_HASH_BEHAVIOUR: merge
+  inventory:
+    links:
+      hosts: ../hosts.yml
+      group_vars: ./group_vars/
 verifier:
   name: testinfra
   options:
-    junit-xml: ../../../../tests/results/uf-rpm-test-results.xml
+    junit-xml: ../../../../tests/results/standalone-systemd-results.xml
